refactor: API, response type, global type 모듈화 (#23)
