+++ src/apis/region/index.ts
@@ -1,0 +1,5 @@
+import { fetchData } from 'apis/fetchData';
+
+export const getRegions = () => {
+  return fetchData('/auth/regions');
+};
