+++ src/components/ProductList/ListItem.tsx
@@ -2,7 +2,7 @@ import { css, styled } from 'styled-components';
 import { getFormattedTimeDifference } from '@utils/time';
 import Icons from '@design/Icons';
 import sample from '../../assets/sample.jpg';
-import { Product } from 'apis/types';
+import { Product } from 'types';
 
 const ListItem: React.FC<Product> = (product) => {
   const {
